+++ src/dotnet/ReSharperPlugin.SpecflowRiderPlugin/QuickFixes/Actions/CreateSpecflowStepFromUsageAction.cs
@@ -1,14 +1,19 @@
+using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Application.UI.Controls;
 using JetBrains.Application.UI.Controls.JetPopupMenu;
+using JetBrains.Collections;
+using JetBrains.Metadata.Reader.API;
+using JetBrains.Metadata.Reader.Impl;
 using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.Impl;
 using JetBrains.ReSharper.Feature.Services.Bulbs;
 using JetBrains.ReSharper.Feature.Services.Navigation.NavigationExtensions;
-using JetBrains.ReSharper.Feature.Services.Resources;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp;
+using JetBrains.ReSharper.Psi.CSharp.Resources;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
+using JetBrains.ReSharper.Psi.Resources;
 using JetBrains.ReSharper.Psi.Transactions;
 using JetBrains.RiderTutorials.Utils;
 using JetBrains.TextControl;
@@ -26,7 +31,9 @@ namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.Actions
         public string Text { get; } = "Create step";
         private readonly SpecflowStepDeclarationReference _reference;
 
-        public CreateSpecflowStepFromUsageAction(SpecflowStepDeclarationReference reference)
+        public CreateSpecflowStepFromUsageAction(
+            SpecflowStepDeclarationReference reference
+        )
         {
             _reference = reference;
         }
@@ -34,12 +41,12 @@ namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.Actions
         private class AvailableBindingClass
         {
             public IPsiSourceFile SourceFile { get; }
-            public string FullClassName { get; }
+            public string ClassClrName { get; }
 
-            public AvailableBindingClass(IPsiSourceFile sourceFile, string fullClassName)
+            public AvailableBindingClass(IPsiSourceFile sourceFile, string classClrName)
             {
                 SourceFile = sourceFile;
-                FullClassName = fullClassName;
+                ClassClrName = classClrName;
             }
         }
 
@@ -66,27 +73,71 @@ namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.Actions
                                        return true;
                                    return false;
                                })
-                        .Select(e => new AvailableBindingClass(e.sourceFile, e.fullClassName));
+                        .Select(e => new AvailableBindingClass(e.sourceFile, e.fullClassName))
+                        .ToList();
+
+                    var filesPerClasses = new OneToSetMap<string, AvailableBindingClass>();
+                    foreach (var availableBindingClass in availableSteps)
+                        filesPerClasses.Add(availableBindingClass.ClassClrName, availableBindingClass);
+
                     // FIXME: Remove full qualifier when not needed
 
-                    menu.ItemKeys.AddRange(availableSteps);
+                    menu.ItemKeys.AddRange(filesPerClasses);
                     menu.DescribeItem.Advise(lifetime, e =>
                                                        {
-                                                           var key = (AvailableBindingClass) e.Key;
-                                                           e.Descriptor.Icon = BulbThemedIcons.RedBulb.Id;
+                                                           var (classClrFullName, availableBindingClasses) = (KeyValuePair<string, ISet<AvailableBindingClass>>) e.Key;
+
+                                                           e.Descriptor.Icon = PsiSymbolsThemedIcons.Class.Id;
                                                            e.Descriptor.Style = MenuItemStyle.Enabled;
-                                                           e.Descriptor.Text = new RichText(key.FullClassName);
+                                                           // FIXME: Use submenu instead of a second popup ? (I did noot find how to get this working)
+                                                           /*if (availableBindingClasses.Count > 1)
+                                                               e.Descriptor.Style |= MenuItemStyle.CanExpand;*/
+
+                                                           var clrTypeName = new ClrTypeName(classClrFullName);
+                                                           e.Descriptor.Text = new RichText(clrTypeName.ShortName, DeclaredElementPresenterTextStyles.ParameterInfo.GetStyle(DeclaredElementPresentationPartKind.Type));
+                                                           e.Descriptor.ShortcutText = clrTypeName.GetNamespaceName();
                                                        });
                     menu.ItemClicked.Advise(lifetime, key =>
                                                       {
-                                                          var targetFile = (AvailableBindingClass) key;
-                                                          AddSpecflowStep(targetFile.SourceFile, _reference.GetStepKind(), _reference.GetStepText());
+                                                          var (_, availableBindingClasses) = (KeyValuePair<string, ISet<AvailableBindingClass>>) key;
+                                                          if (availableBindingClasses.Count == 1)
+                                                          {
+                                                              var availableBindingClass = availableBindingClasses.First();
+                                                              AddSpecflowStep(availableBindingClass.SourceFile, availableBindingClass.ClassClrName, _reference.GetStepKind(), _reference.GetStepText());
+                                                          }
+                                                          else
+                                                          {
+                                                              OpenFileSelectionModal(jetPopupMenus, textControl, availableBindingClasses, _reference.GetStepKind(), _reference.GetStepText());
+                                                          }
                                                       });
                     menu.PopupWindowContextSource = textControl.PopupWindowContextFactory.ForCaret();
                 });
         }
 
-        private static void AddSpecflowStep(IPsiSourceFile targetFile, GherkinStepKind stepKind, string stepText)
+        private void OpenFileSelectionModal(JetPopupMenus jetPopupMenus, ITextControl textControl, ISet<AvailableBindingClass> availableBindingClasses, GherkinStepKind getStepKind, string getStepText)
+        {
+            jetPopupMenus.ShowModal(JetPopupMenu.ShowWhen.AutoExecuteIfSingleEnabledItem,
+                (lifetime, menu) =>
+                {
+                    menu.Caption.Value = WindowlessControlAutomation.Create("Where to create the step ?");
+                    menu.ItemKeys.AddRange(availableBindingClasses);
+                    menu.DescribeItem.Advise(lifetime, e =>
+                                                       {
+                                                           var key = (AvailableBindingClass) e.Key;
+                                                           e.Descriptor.Icon = PsiCSharpThemedIcons.Csharp.Id;
+                                                           e.Descriptor.Style = MenuItemStyle.Enabled;
+                                                           e.Descriptor.Text = new RichText(key.SourceFile.DisplayName);
+                                                       });
+                    menu.ItemClicked.Advise(lifetime, e =>
+                                                      {
+                                                          var availableBindingClass = (AvailableBindingClass) e;
+                                                          AddSpecflowStep(availableBindingClass.SourceFile, availableBindingClass.ClassClrName, getStepKind, getStepText);
+                                                      });
+                    menu.PopupWindowContextSource = textControl.PopupWindowContextFactory.ForCaret();
+                });
+        }
+
+        private static void AddSpecflowStep(IPsiSourceFile targetFile, string classClrName, GherkinStepKind stepKind, string stepText)
         {
             var cSharpFile = targetFile.GetProject().GetCSharpFile(targetFile.DisplayName.Substring(targetFile.DisplayName.LastIndexOf('>') + 2));
             if (cSharpFile == null)
@@ -95,6 +146,8 @@ namespace ReSharperPlugin.SpecflowRiderPlugin.QuickFixes.Actions
             foreach (var type in cSharpFile.GetChildrenInSubtrees<IClassDeclaration>())
             {
                 if (!(type is IClassDeclaration classDeclaration))
+                    continue;
+                if (classDeclaration.CLRName != classClrName)
                     continue;
                 if (classDeclaration.DeclaredElement?.GetAttributeInstances(AttributesSource.Self).All(x => x.GetAttributeType().GetClrName().FullName != "TechTalk.SpecFlow.Binding") != true)
                     continue;